requirements.txt
attrs==23.2.0
certifi==2024.7.4
defusedxml==0.7.1
et-xmlfile==1.1.0
h11==0.14.0
idna==3.7
numpy==2.0.0
odfpy==1.4.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
selenium==4.23.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
